# server/docs/swagger.yaml
definitions:
...
  request.ItemQuantityRequest:
    properties:
      quantity:
        type: integer
      sku_id:
        type: string
    type: object
...
  request.QuantityRequest:
    properties:
      quantity:
        type: integer
    type: object
...
  request.SkuIdRequest:
    properties:
      skuId:
        type: string
    required:
    - skuId
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.8.1
paths:
...
  /cart:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 現在のカート内容（商品リスト、合計情報）を取得
      security:
      - ApiKeyAuth: []
      summary: 現在のカート内容（商品リスト、合計情報）を取得
      tags:
      - Cart
  /cart/items:
    post:
      consumes:
      - application/json
      parameters:
      - description: SKU ID,Quantity
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ItemQuantityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 指定したSKUと数量をカートに追加（または更新）
      security:
      - ApiKeyAuth: []
      summary: 指定したSKUと数量をカートに追加（または更新）
      tags:
      - Cart
  /cart/items/{skuId}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: sku id
        in: path
        name: skuId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: カート内の指定SKUを削除する
      security:
      - ApiKeyAuth: []
      summary: カート内の指定SKUを削除する
      tags:
      - Cart
    put:
      consumes:
      - application/json
      parameters:
      - description: sku id
        in: path
        name: skuId
        required: true
        type: string
      - description: Quantity
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.QuantityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: カート内の指定SKUの数量を変更する
      security:
      - ApiKeyAuth: []
      summary: カート内の指定SKUの数量を変更する
      tags:
      - Cart
...
  /favorites/skus:
    get:
      consumes:
      - application/json
      parameters:
      - description: ページ番号 (1始まり)
        in: query
        name: page
        type: integer
      - description: 1ページあたりのレビュー件数
        in: query
        name: limit
        type: integer
      - description: ソート順序('newest', 'oldest'のいずれかを指定
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 自身のお気に入りSKUリストを取得する
      security:
      - ApiKeyAuth: []
      summary: 自身のお気に入りSKUリストを取得する
      tags:
      - Product
    post:
      consumes:
      - application/json
      parameters:
      - description: SKU ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.SkuIdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 指定したSKUをお気に入りに追加する
      security:
      - ApiKeyAuth: []
      summary: 指定したSKUをお気に入りに追加する
      tags:
      - Product
  /favorites/skus/{skuId}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: sku id
        in: path
        name: skuId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 指定したSKUをお気に入りから削除する
      security:
      - ApiKeyAuth: []
      summary: 指定したSKUをお気に入りから削除する
      tags:
      - Product
...
  /history/viewed-skus:
    get:
      consumes:
      - application/json
      parameters:
      - description: ページ番号 (1始まり)
        in: query
        name: page
        type: integer
      - description: 1ページあたりのレビュー件数
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 自身の最近チェックした商品（SKU）リストを取得する
      security:
      - ApiKeyAuth: []
      summary: 自身の最近チェックした商品（SKU）リストを取得する
      tags:
      - Product
    post:
      consumes:
      - application/json
      parameters:
      - description: SKU ID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.SkuIdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 閲覧したSKU（代表SKU）を履歴に記録する
      security:
      - ApiKeyAuth: []
      summary: 閲覧したSKU（代表SKU）を履歴に記録する
      tags:
      - Product
...
  /products/getCoordinates/{productCode}/coordinates:
    get:
      consumes:
      - application/json
      parameters:
      - description: Product Code
        in: path
        name: productCode
        required: true
        type: string
      - description: 表示するコーディネートセット概要の最大件数
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 商品コーディネート概要リスト取得
      summary: 商品コーディネート概要リスト取得
      tags:
      - Product
  /products/getImage/{skuId}/images:
    get:
      consumes:
      - application/json
      parameters:
      - description: Sku id
        in: path
        name: skuId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: SKU画像リスト取得
      summary: SKU画像リスト取得
      tags:
      - Product
  /products/getProductInfo/{productCode}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Product Code
        in: path
        name: productCode
        required: true
        type: string
      - description: SKU ID
        in: query
        name: sku_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 商品の基本情報と、リクエストで特定されたSKUに関する詳細情報を返す
      security:
      - ApiKeyAuth: []
      summary: 商品情報取得
      tags:
      - Product
  /products/getQA/{productCode}/qa:
    get:
      consumes:
      - application/json
      parameters:
      - description: Product Code
        in: path
        name: productCode
        required: true
        type: string
      - description: ページ番号 (1始まり)
        in: query
        name: page
        type: integer
      - description: 1ページあたりのQ&A件数
        in: query
        name: limit
        type: integer
      - description: ソート順序('newest', 'oldest', 'most_helpful' のいずれかを指定
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 紐づく承認済みの質問とそれに対する店舗回答のリストを、ページネーションおよびソート機能付きで取得する
      summary: Q&A一覧取得
      tags:
      - Product
  /products/getRelated/{productCode}/related:
    get:
      consumes:
      - application/json
      parameters:
      - description: Product Code
        in: path
        name: productCode
        required: true
        type: string
      - description: 表示する関連商品の最大件数
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 関連商品リスト取得
      summary: 関連商品リスト取得
      tags:
      - Product
  /products/getReviews/{productCode}/reviews:
    get:
      consumes:
      - application/json
      parameters:
      - description: Product Code
        in: path
        name: productCode
        required: true
        type: string
      - description: ページ番号 (1始まり)
        in: query
        name: page
        type: integer
      - description: 1ページあたりのレビュー件数
        in: query
        name: limit
        type: integer
      - description: ソート順序('newest', 'oldest', 'highest_rating', 'lowest_rating',
          'most_helpful' のいずれかを指定
        in: query
        name: sort
        type: string
      - description: '指定した評価（星の数、例: 5）'
        in: query
        name: rating
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: レビューリスト、集計情報、ページネーション情報を返す
      security:
      - ApiKeyAuth: []
      summary: レビュー一覧取得
      tags:
      - Product
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"
...